0.7.2.0 (Using non-targetable ability order)

+++ Scripts/AutoAttackProjectile.cpp
@@ -64,7 +64,7 @@ void AutoAttackProjectile::Update(CallbackRecArgs_LUpd args) {
 	else {
 		Move_Global(dir * Speed);
 		if (!LockRotation)
-			SetGlobalRotation(DirectionToAngle(dir));
+			SetGlobalRotation((float)DirectionToAngle(dir));
 	}
 }
 const AttackHitInfo& AutoAttackProjectile::GetAttInfo(IAttackableObj& target) const {
